+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # 2.9.62
 
+* BREAKING CHANGE: UPDATE TO D&D BEYOND - there is a chance that some custom items may no longer update correctly to D&D Beyond. If an error UI message pops up, ensure that the items on DDB reflect your character in Foundry and reimport.
 * Some actors could choke during import if using Item Containers/Collections module.
 
 # 2.9.61